Added Null Userid error handling

+++ apis/profileAPI.ts
@@ -36,7 +36,11 @@ export const changePasswordRequest = async (userId: string | null, data: ChangeP
   if (!res.ok) throw new Error("Kunne ikke opdatere adgangskode");
 };
 
-export const deleteUserRequest = async (userId: string, data: DeleteUserDTO) => {
+export const deleteUserRequest = async (userId: string | null, data: DeleteUserDTO) => {
+  if (userId === null) {
+    throw new Error("FATAL FEJL: Bruger-ID er ikke korrekt initialiseret i din session.");
+  }
+
   const res = await fetch(`${BASE_URL}/users/${userId}`, {
     method: "DELETE",
     body: JSON.stringify(data),
